+++ geographic_services/settings/base.py
@@ -68,6 +68,21 @@ SPECTACULAR_SETTINGS = {
     'SERVE_INCLUDE_SCHEMA': False
 }
 
+CACHE_TTL = 60 * 1
+CACHES = {
+    'default': {
+        'BACKEND': 'django_redis.cache.RedisCache',
+        'LOCATION': 'redis://127.0.0.1:6379',
+        'OPTIONS': {
+            'CLIENT_CLASS': 'django_redis.client.DefaultClient',
+        }
+    }
+}
+SESSION_ENGINE = 'django.contrib.sessions.backends.cache'
+SESSION_CACHE_ALIAS = 'default'
+PROVIDER_CACHE_TTL = 60 * 5
+SERVICE_AREA_CACHE_TTL = 60 * 5
+
 
 ROOT_URLCONF = 'geographic_services.urls'