+++ geographic_services/settings/test.py
@@ -4,3 +4,9 @@ from .base import mongoengine
 mongoengine.disconnect()
 
 DEBUG = True
+
+CACHES = {
+    'default': {
+        'BACKEND': 'django.core.cache.backends.dummy.DummyCache',
+    }
+}