+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
    - Service area app module
    - Endpoint to get service areas based on latitude and longitude
+   - Cache configuration and decorators
 
 ## [0.2.0] - 2022-12-04 13:34